feat(stencil): use urn as schemaid
